Rename modules
Rename modules with better name and move the arduino code to its own folder
