+++ devicecommunication/CommunicationMain.py
