+++ gcodeextractor/tests/resources/expected/text1.txt
