+++ gcodeextractor/tests/resources/expected/text2.txt
