+++ gcodeextractor/tests/resources/output/text1.txt
@@ -1,0 +1,5 @@
+This
+is
+a
+
+test
