+++ gcodeextractor/tests/resources/output/text2.txt
@@ -1,0 +1,5 @@
+G28
+G90
+G0 Z3
+
+G0 X15 Y45
