+++ gcodeextractor/tests/resources/raw/test1.pbm
