+++ gcodeextractor/tests/resources/raw/test100x100.pbm
