+++ gcodeextractor/tests/resources/raw/test1ascii.pbm
